[frontend] fix `askJobImport` import/no-cycle (#9760)
